add edge resource estimation

+++ src/fpga/microblossom/util/Vivado.scala
@@ -20,105 +20,14 @@ object VivadoTarget {
 
 }
 
-case class TimingReportPathNode(
-    var increment: Double,
-    var accumulate: Double,
-    var location: String = null,
-    var delayName: String = null,
-    var delayAux: String = null,
-    var netlistResource: String = null
-)
-
-case class TimingReportPath(
-    var slack: String = null,
-    var source: String = null,
-    var destination: String = null,
-    var dataPathDelay: String = null,
-    var logicLevels: String = null,
-    var details: mutable.ArrayBuffer[TimingReportPathNode] = null
-)
-
-case class TimingReport(filepath: String) {
-  val paths = mutable.ArrayBuffer[TimingReportPath]()
-
-  val source = Source.fromFile(filepath)
-  try {
-    var path: TimingReportPath = null
-    var capturePathNode = false
-    val pathNodeRegex: Regex = """(\S+) \(([^)]*)\)\s+(\d+.\d+)\s+(\d+.\d+)\s+\w?\s+(\S+)""".r
-    for (originalLine <- source.getLines()) {
-      val line = originalLine.trim
-      if (line.startsWith("Slack:")) {
-        path = TimingReportPath()
-        path.slack = line.split(":")(1).trim
-      } else if (line.startsWith("Source:")) {
-        path.source = line.split(":")(1).trim
-      } else if (line.startsWith("Destination:")) {
-        path.destination = line.split(":")(1).trim
-      } else if (line.startsWith("Data Path Delay:")) {
-        path.dataPathDelay = line.split(":")(1).trim
-      } else if (line.startsWith("Logic Levels:")) {
-        path.logicLevels = line.split(":")(1).trim
-      } else if (path != null && line.startsWith("-------------------------")) {
-        if (path.details == null) {
-          path.details = mutable.ArrayBuffer()
-          capturePathNode = true
-        } else {
-          paths.append(path)
-          path = null
-          capturePathNode = false
-        }
-      } else {
-        if (capturePathNode) {
-          line match {
-            case pathNodeRegex(delayName, delayAux, increment, accumulate, netlistResource) =>
-              path.details.append(
-                TimingReportPathNode(
-                  delayName = delayName,
-                  delayAux = delayAux,
-                  increment = increment.toDouble,
-                  accumulate = accumulate.toDouble,
-                  netlistResource = netlistResource
-                )
-              )
-            case _ => // no nothing
-          }
-        }
-      }
-    }
-  } finally {
-    source.close()
-  }
-
-  /** analyze delay excluding IBUF and OBUF and their associated net delay */
-  def getPathDelaysExcludingIO(): List[Double] = {
-    val delays = mutable.ArrayBuffer[Double]()
-    for (path <- paths) {
-      var startingTime = path.details(0).accumulate
-      var endingTime = path.details(path.details.length - 1).accumulate
-      for ((pathNode, index) <- path.details.zipWithIndex) {
-        if (pathNode.delayName == "IBUF") {
-          startingTime = path.details(index + 1).accumulate
-        }
-        if (pathNode.delayName == "OBUF") {
-          endingTime = path.details(index - 2).accumulate
-        }
-      }
-      delays.append(endingTime - startingTime)
-    }
-    delays.toList
-  }
-
-  def getPathDelaysExcludingIOWorst(): Double = getPathDelaysExcludingIO.max
-
-}
-
 object Vivado {
 
   /** when `useImpl` is true, run implementation to get a better estimation of timing */
-  def reportTiming[T <: Component](component: => T, useImpl: Boolean = false): TimingReport = {
-
-    // return TimingReport("gen/tmp/gwlkxvkqjF/timing.txt")
+  def createProject[T <: Component](
+      component: => T,
+      useImpl: Boolean = false,
+      removeVivadoProj: Boolean = false
+  ): String = {
 
     val projectName = Random.alphanumeric.filter(_.isLetter).take(10).mkString
     val targetDirectory = s"gen/tmp/$projectName"
@@ -183,16 +92,34 @@ set_output_delay 0 -clock virt_clk [all_outputs]
 $runJobScript
 
 report_timing -from [all_inputs] -to [all_outputs] -nworst 10 -file ./timing.txt
+report_utilization -file ./resource.txt
 """
-    Files.write(Paths.get(s"$targetDirectory/reportTiming.tcl"), script.getBytes(StandardCharsets.UTF_8))
+    Files.write(Paths.get(s"$targetDirectory/create_project.tcl"), script.getBytes(StandardCharsets.UTF_8))
     // run the TCL script to create vivado project
-    val command = "vivado -mode batch -source reportTiming.tcl"
+    val command = "vivado -mode batch -source create_project.tcl"
     val folder = new java.io.File(targetDirectory)
     val stdoutFile = new java.io.File(s"$targetDirectory/output.txt")
     // val output = (Process(command, folder) #> stdoutFile).!!
     val output = Process(command, folder).!!
     // remove temporary vivado project when the command succeeded
-    Process(s"rm -rf $projectName", folder).!!
+    if (removeVivadoProj) {
+      Process(s"rm -rf $projectName", folder).!!
+    }
+
+    targetDirectory
+
+  }
+
+  def reportTiming[T <: Component](component: => T, useImpl: Boolean = false): TimingReport = {
+    // return TimingReport("gen/tmp/gwlkxvkqjF/timing.txt")
+    val targetDirectory = createProject(component, useImpl = useImpl)
     TimingReport(s"$targetDirectory/timing.txt")
+  }
+
+  /** when `useImpl` is true, run implementation to get a better estimation of timing */
+  def reportResource[T <: Component](component: => T, useImpl: Boolean = false): ResourceReport = {
+    // return ResourceReport("gen/tmp/eLrnekNLbJ/resource.txt")
+    val targetDirectory = createProject(component, useImpl = useImpl)
+    ResourceReport(s"$targetDirectory/resource.txt")
   }
 }
